+++ api/src/kline.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 struct Kline {
     time: u32,
-    open: f32,
-    close: f32,
-    high: f32,
-    low: f32,
-    volume: f32,
+    open: f64,
+    close: f64,
+    high: f64,
+    low: f64,
+    volume: f64,
 }
